backend/requirements.txt
tensorflow==2.0.0
flask==1.1.1
google-cloud-storage==1.23.0